+++ app-modules/integration-open-ai/src/Services/BaseOpenAiService.php
@@ -40,21 +40,16 @@ use Closure;
 use Generator;
 use AdvisingApp\Ai\Models\AiThread;
 use AdvisingApp\Ai\Models\AiMessage;
-use Illuminate\Support\Facades\Http;
 use OpenAI\Contracts\ClientContract;
 use AdvisingApp\Ai\Models\AiAssistant;
 use AdvisingApp\Ai\Settings\AiSettings;
-use AdvisingApp\Ai\Models\AiMessageFile;
 use OpenAI\Responses\Threads\ThreadResponse;
 use AdvisingApp\Ai\Services\Contracts\AiService;
 use OpenAI\Responses\Threads\Runs\ThreadRunResponse;
 use AdvisingApp\Ai\Exceptions\MessageResponseException;
 use AdvisingApp\Ai\Services\Concerns\HasAiServiceHelpers;
 use AdvisingApp\Ai\Exceptions\MessageResponseTimeoutException;
-use AdvisingApp\Ai\DataTransferObjects\Files\FilesDataTransferObject;
 use AdvisingApp\Ai\DataTransferObjects\Threads\ThreadsDataTransferObject;
-use AdvisingApp\Ai\DataTransferObjects\VectorStores\VectorStoresDataTransferObject;
-use AdvisingApp\Ai\DataTransferObjects\VectorStoreFiles\VectorStoreFilesDataTransferObject;
 
 abstract class BaseOpenAiService implements AiService
 {
@@ -416,78 +411,6 @@ abstract class BaseOpenAiService implements AiService
     public function supportsFileUploads(): bool
     {
         return false;
-    }
-
-    public function retrieveFile(AiMessageFile $file): FilesDataTransferObject
-    {
-        $response = $this->client->files()->retrieve($file->file_id);
-
-        return FilesDataTransferObject::from([
-            'id' => $response->id,
-            'name' => $response->filename,
-            'status' => $response->status,
-        ]);
-    }
-
-    public function createVectorStore(array $parameters): VectorStoresDataTransferObject
-    {
-        $response = $this->client->vectorStores()->create($parameters);
-
-        return VectorStoresDataTransferObject::from([
-            'id' => $response->id,
-            'name' => $response->name,
-            'fileCounts' => get_object_vars($response->fileCounts),
-            'status' => $response->status,
-            'expiresAt' => $response->expiresAt,
-        ]);
-    }
-
-    public function retrieveVectorStore(string $vectorStoreId): VectorStoresDataTransferObject
-    {
-        $response = $this->client->vectorStores()->retrieve($vectorStoreId);
-
-        return VectorStoresDataTransferObject::from([
-            'id' => $response->id,
-            'name' => $response->name,
-            'fileCounts' => get_object_vars($response->fileCounts),
-            'status' => $response->status,
-            'expiresAt' => $response->expiresAt,
-        ]);
-    }
-
-    public function modifyVectorStore(string $vectorStoreId, array $parameters): void
-    {
-        $response = $this->client->vectorStores()->modify($vectorStoreId, $parameters);
-    }
-
-    /**
-     * The `openai-php/client` does not current work with the `GET /vector_stores/{vectorStoreId}/files` endpoint
-     * for Azure Open AI. This is due to the expectation of a `chunking_strategy` key in the response, which Azure
-     * does not provide. An issue has been opened, but this request needs to happen without the client for now.
-     */
-    public function retrieveVectorStoreFiles(AiThread $thread, string $vectorStoreId, array $params): VectorStoreFilesDataTransferObject
-    {
-        $service = $thread->assistant->model->getService();
-
-        $response = Http::withHeaders([
-            'api-key' => $service->getApiKey(),
-            'OpenAI-Beta' => 'assistants=v2',
-            'Content-Type' => 'application/json',
-        ])
-            ->withQueryParameters(
-                $params,
-            )
-            ->get($service->getDeployment() . '/vector_stores/' . $vectorStoreId . '/files', [
-                'api-version' => $service->getApiVersion(),
-            ]);
-
-        return VectorStoreFilesDataTransferObject::from([
-            'object' => $response->json()['object'],
-            'data' => $response->json()['data'],
-            'firstId' => $response->json()['first_id'],
-            'lastId' => $response->json()['last_id'],
-            'hasMore' => $response->json()['has_more'],
-        ]);
     }
 
     protected function awaitThreadRunCompletion(ThreadRunResponse $threadRunResponse): void
